+++ internal/acctest/oauthauthserversettingsscopescommonscopes/oauth_auth_server_settings_scopes_common_scopes_test.go
@@ -69,11 +69,10 @@ func TestAccOauthAuthServerSettingsScopesCommonScopes(t *testing.T) {
 func testAccOauthAuthServerSettingsScopesCommonScopes(resourceName string, resourceModel oauthAuthServerSettingsScopesCommonScopesResourceModel) string {
 	return fmt.Sprintf(`
 resource "pingfederate_oauth_auth_server_settings_scopes_common_scopes" "%[1]s" {
-  dynamic     = %[3]t
-  description = "%[4]s"
-  name        = "%[5]s"
+  dynamic     = %[2]t
+  description = "%[3]s"
+  name        = "%[4]s"
 }`, resourceName,
-		resourceModel.id,
 		resourceModel.dynamic,
 		resourceModel.description,
 		resourceModel.name,